Merge in CHMS/chimes_lsq from ~LINDSEY11/chimes_lsq-fork:public_release to public_release

* commit 'd5504b7df2b9e39b6d523571f5842aacf1bc67bf': (31 commits)
  Resolve merge conflicts and fix text suite. + Many test suite runs failed due to trivial text header formatting + Some tests fail due to randomized placements of penalty kick-in warnings   - Penalty warnings are otherwise identical across correct and current outputs + Some tests did not have A matrix split + MD test suite/generator did not reliably recompile code + TATB lsq fm_setup.in orders did not match correct output orders + Fixed changes that broke AL driver functionality + Changed "SKIP_FRAMES" chimes_lsq input var to "SKPFRMS" since all input vars are 7 chars + Fixed chimes_lsq test suite call to chimes_lsq.py, which used py2 instead of py3
  Added new tests of Ti with ortho and non-ortho unit cells.
  Fixed a bug in chimes_lsq when an orthogonal cell was input as non-orthogonal. Added a test of split output files, dlars, and orthogonal cell bug.
  Fix for a bug in the calculation of neighbors for non-orthorhombic unit cells. Added a test of molecular dynamics with a triclinic unit cell. Added tests to enforce LAMMPS unit cell conventions on xyz input.
  Implemented a full restart capability for dlars.
  Added a new file for correct output.
  Committing all changes to Larry's chimes_lsq.dlars repo.
  Added an mpi barrier to prevent timeouts.
  Update to dlars test Makefile to reflect change in default normalization.
  Documented skip_frames features in chimes_lsq.
  Updated DLARS test output.
  Documented feature weights.
  Added support for feature weights. Changed diagnostic output.
  Changes to whitespace in header files.
  Removed comment from CMakeLists.txt.
  git will ignore forceout.txt differences.
  Added a new ignore phrase to compare.pl.
  Fix error in a comment.
  Removed compilation of dlars in the test suite scripts.  It is handled by install.sh.
  Updated lsq test output.  Removed a non-existent lsq test (split_files).
  ...

+++ install.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+echo ""
+echo "***"
+echo "*** Executing chimes_lsq install.sh"
+echo "***" 
+echo ""
+
 # Builds all relevant chimes_calculator executables/library files
 # Run with:
 # ./install.sh
@@ -12,15 +18,18 @@ VERBO=${3-1}  # Verbosity set to 1 by default, 0 gives minimal output, 3 gives m
 DOMPI=${4-1}  # Compile with MPI support by default
 
 
-echo "Performing a fresh install"
-
+echo "Attempting to perform a fresh install"
 echo "Imports directory will be deleted and re-cloned/installed. Proceed? (y/n)"
 read PROCEED
 if [[ "$PROCEED" == "n" ]] ; then
-	exit 0
+    echo 'Will use pre-existing imports directory'
+else
+    echo "Are you sure? Imports directory will be deleted! (y/n)"
+    read PROCEED
+    if [[ "$PROCEED" == "y" ]] ; then
+        rm -rf imports
+    fi
 fi
-rm -rf imports
-
 
 # Determine computing environment
 
@@ -33,13 +42,11 @@ read IS_LC
 ICC=`which g++`
 
 if [[ "$IS_LC" == "y" ]] ; then
-
-	module load intel/18.0.1
-	module load impi/2018.0
-	module load cmake/3.21.1
-	module load mkl/2020.0	
-	
-	ICC=`which icc`
+    module load cmake/3.21.1
+    module load intel/18.0.1
+    module load impi/2018.0
+    
+    ICC=`which icc`
 fi
 
 MPI=`which mpicxx` # /usr/tce/packages/mvapich2/mvapich2-2.3-intel-18.0.1/bin/mpicxx
@@ -47,14 +54,16 @@ MPI=`which mpicxx` # /usr/tce/packages/mvapich2/mvapich2-2.3-intel-18.0.1/bin/mp
 
 # Grab and install dependencies
 
-./clone-all.sh
+if [[ ! -d imports ]] ; then
+    ./clone-all.sh 1 $IS_LC
+fi
 
 # Compile dlars
 
 if [[ "$IS_LC" == "y" ]] ; then
-	cd contrib/dlars/src
-	make
-	cd - 1&>/dev/null
+    cd contrib/dlars/src
+    make
+    cd - 1&>/dev/null
 fi
 
 # Compile molanal
@@ -81,28 +90,28 @@ if [ ! -z $PREFX ] ; then
 fi
 
 if [ $DEBUG -eq 1 ] ;then
-	my_flags="${my_flags} -Wall -DCMAKE_BUILD_TYPE=Debug"
+    my_flags="${my_flags} -Wall -DCMAKE_BUILD_TYPE=Debug"
 else
-	my_flags="${my_flags} -DCMAKE_BUILD_TYPE=Release"
+    my_flags="${my_flags} -DCMAKE_BUILD_TYPE=Release"
 fi
 
 if   [ $VERBO -eq 0 ] ;then
         my_flags="${my_flags} -DVERBOSITY=0" 
-	my_flags="${my_flags} -DDEBUG_CHEBY=0" 
+    my_flags="${my_flags} -DDEBUG_CHEBY=0" 
 elif [ $VERBO -eq 1 ] ; then
         my_flags="${my_flags} -DVERBOSITY=1" 
-	my_flags="${my_flags} -DDEBUG_CHEBY=0" 
+    my_flags="${my_flags} -DDEBUG_CHEBY=0" 
 elif [ $VERBO -eq 2 ] ; then
         my_flags="${my_flags} -DVERBOSITY=0" 
-	my_flags="${my_flags} -DDEBUG_CHEBY=1" 	
+    my_flags="${my_flags} -DDEBUG_CHEBY=1"     
 elif [ $VERBO -eq 3 ] ; then
         my_flags="${my_flags} -DVERBOSITY=1" 
-	my_flags="${my_flags} -DDEBUG_CHEBY=1" 	
+    my_flags="${my_flags} -DDEBUG_CHEBY=1"     
 fi
 
 if [ $DOMPI -eq 1 ] ;then
         my_flags="${my_flags} -DUSE_MPI=1" 
-	my_flags="${my_flags} -DMPI_CXX_COMPILER=${MPI}"
+    my_flags="${my_flags} -DMPI_CXX_COMPILER=${MPI}"
 else
         my_flags="${my_flags} -DUSE_MPI=0" 
 fi
@@ -119,30 +128,32 @@ cp ../src/chimes_lsq.py .
 cp ../src/post_proc_chimes_lsq.py .
 
 if [ $DOMPI -eq 1 ] ;then
+ 
+ # Create some executables for the ALD
 
-	# Create some executables for the ALD
-	
-	cp chimes_md chimes_md-mpi
-	cp chimes_lsq chimes_lsq.tmp
-	
-	my_flags=`echo $my_flags | awk '{for(i=1;i<=NF; i++){if($i~"DUSE_MPI=1"){$i="-DUSE_MPI=0"}}{print}}'`
-	
-	cmake $my_flags ..
-	make
-	
-	cp chimes_md chimes_md-serial
-	cp chimes_md-mpi chimes_md
-	mv chimes_lsq.tmp chimes_lsq
+    cp chimes_md chimes_md-mpi
+    cp chimes_lsq chimes_lsq.tmp
+
+    my_flags=`echo $my_flags | awk '{for(i=1;i<=NF; i++){if($i~"DUSE_MPI=1"){$i="-DUSE_MPI=0"}}{print}}'`
+
+    cmake $my_flags ..
+    make
+
+    cp chimes_md chimes_md-serial
+    cp chimes_md-mpi chimes_md
+    mv chimes_lsq.tmp chimes_lsq
 fi
-	
-	
-
 
 if [ ! -z $PREFX ] ; then
         make install
-	cp src/chimes_lsq.py $PREFX
-	cp src/post_proc_chimes_lsq.py $PREFX
+        cp src/chimes_lsq.py $PREFX
+    cp src/post_proc_chimes_lsq.py $PREFX
 fi
 
 cd ..
 
+echo ""
+echo "***"
+echo "*** chimes_lsq install complete"
+echo "***"
+echo ""
